DeskBooking.Core/Processor/DeskBookingRequestProcessor.cs
using DeskBooking.Core.Domain;
using DeskBooking.Core.Interfaces;

namespace DeskBooking.Core.Processor
{
    public class DeskBookingRequestProcessor
    {
        private IDeskBookingRepository _deskBookingRepository;
        private IDeskRepository _deskRepository;

        public DeskBookingRequestProcessor(IDeskBookingRepository deskBookingRepository, IDeskRepository deskRepository)
        {
            _deskBookingRepository = deskBookingRepository;
            _deskRepository = deskRepository;
        }

        public DeskBookingResult BookDesk(DeskBookingRequest request)
        {
            if (request == null)
            {
                throw new ArgumentNullException(nameof(request));
            }

            var availableDesks = _deskRepository.GetAvailableDesks(request.BookingDate);
            if (availableDesks.Count() > 0)
            {
                _deskBookingRepository.Save(Create<DeskBookingModel>(request));
            }

            return Create<DeskBookingResult>(request);
        }

        private static T Create<T>(DeskBookingRequest request) where T : DeskBookingBase, new()
        {
            return new T
            {
                FirstName = request.FirstName,
                LastName = request.LastName,
                Email = request.Email,
                BookingDate = request.BookingDate,
            };
        }


    }
}